Update README.md
tried to include a button as part of a list element but it was triggering a drag event when clicked. fished around in the source code and saw you had a noDragClass to prevent. I didn't find it in your documentation so I added it. Great job on the plugin!