Added Support for Max Items in List
Support option for denying a user to limit the number of items in a nestable list. This concerns itself only with the destination (drop list) item count. The comparison for >=) is in the case that more than the item limit was rendered to the list (i.e. even though you limited nestable to 3 items, the server side template rendered 5 items). I realize this could be more concise and have better performance, but is presented for readability.